pyproject.toml
[project]
name = "llm-react-benchmark"
version = "0.1.0"
description = "Benchmarking React programming tasks using LLMs with scoring via Claude"
authors = ["Roman Matena <info@romanmatena.cz>"]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
